Revert "Enable simpledrm support"
This reverts commit 16ebb3196a62804c351ce3dc25b09df7a98bd716.

Simpledrm could only function with a simplefb, which conflicts with
traditional framebuffer support (exported as /dev/fbX). However, the
commit turns SIMPLEDRM to m, which means a system without a dedicated
DRM device won't show anything on the display even if a firmware-
initialized framebuffer is available until simpledrm is loaded.

Unluckily, simpledrm carries no MODULE_DEVICE_TABLE for simplefb, which
acts as a platform device. This makes systems without a dedicated
DRM device impossible to react when booting eweOS. It's known that
simpledrm sometimes causes strange performance degradation when used
together with a dedicated DRM device, so building it into the kernel
isn't an option, either.

Let's revert the change until coming up with a better solution to have
KMS available on framebuffer-only systems.

Signed-off-by: Yao Zi <ziyao@disroot.org>